Move some parser recovery methods to diagnostics
